+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixed: `as.data.table()` method for `featureGroups`: `normConcToTox` argument was ignored (not for `featureGroupsScreening`)
 * Fixed: `report()`: don't show (empty) annotations column in feature data if no feature annotations are present
+* Fixed: featureGroups annotations slot is properly updated if all features are removed from a set
 
 # patRoon 2.3.1